backend/requirements.txt
fastapi
supabase
uvicorn
python-dotenv
google-generativeai